requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.1
django-admin-display==1.3.0
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
flake8==5.0.4
flake8-annotations==2.9.1
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
iniconfig==1.1.1
install==1.3.5
mccabe==0.7.0
mysqlclient==2.1.1
packaging==21.3
pluggy==1.0.0
promise==2.3
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
strawberry-django-jwt==0.2.2
strawberry-django-optimizer==0.1.0
strawberry-django-plus==1.28.2
strawberry-graphql==0.133.5
strawberry-graphql-django==0.5.3
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
